ignore the file_rename.py itself when rename files

+++ utils/file_rename.py
@@ -17,6 +17,10 @@ frename.change_postfix(old_postfix='png',new_postfix='jpg') a.png --> a.jpg
 frename.add_postfix(self,new_postfix='txt') a --> a.txt
 frename.add_prefix(self,prefix='xx',before=True) a.txt --> test_a.txt/ a_test.txt
 frename.rename_number(self,prefix='xx') a.txt --> test_001.txt
+
+changelist:
+1. init version
+2. ignore the program itself, not to process file_rename.py
 '''
 class FileRename(object):
     def __init__(self,work_dir='.'):
@@ -27,6 +31,7 @@ class FileRename(object):
                 if self.mkdir_temp(work_dir):
                     self.file_process_cnt=0
                     self.inited=True
+                    self.fn_myself=os.path.basename(os.path.abspath(__file__))
             else:
                 print '"%s" is not a dir' % work_dir
         else:
@@ -47,6 +52,10 @@ class FileRename(object):
         if not self.inited:
             print 'not inited,quit'
             sys.exit(1)
+    def is_myself(self,fn):
+        if fn == self.fn_myself:
+            return True
+        return False
     def change_postfix(self,old_postfix='png',new_postfix='jpg'):
         self.check_inited()
         work_dir=self.work_dir
@@ -66,7 +75,7 @@ class FileRename(object):
         work_dir=self.work_dir
         cnt=self.file_process_cnt=0
         for fn in os.listdir(work_dir):
-            if os.path.isfile(os.path.join(work_dir,fn)):
+            if os.path.isfile(os.path.join(work_dir,fn)) and not self.is_myself(fn):
                 file_root=os.path.splitext(fn)[0]
                 file_new=fn+'.'+new_postfix
                 #print fn,file_new
@@ -79,7 +88,7 @@ class FileRename(object):
         work_dir=self.work_dir
         cnt=self.file_process_cnt=0
         for fn in os.listdir(work_dir):
-            if os.path.isfile(os.path.join(work_dir,fn)):
+            if os.path.isfile(os.path.join(work_dir,fn)) and not self.is_myself(fn):
                 file_root,file_ext=os.path.splitext(fn)
                 if before:
                     file_new=prefix+'_'+file_root+file_ext
@@ -95,9 +104,9 @@ class FileRename(object):
         work_dir=self.work_dir
         cnt=self.file_process_cnt=0
         for fn in os.listdir(work_dir):
-            if os.path.isfile(os.path.join(work_dir,fn)):
+            if os.path.isfile(os.path.join(work_dir,fn)) and not self.is_myself(fn):
                 file_root,file_ext=os.path.splitext(fn)
-                file_new=prefix+'_'+'{0:03d}'.format(cnt)+file_ext
+                file_new=prefix+'_'+'{0:04d}'.format(cnt)+file_ext
                 print fn,file_new
                 os.rename(os.path.join(work_dir,fn),os.path.join(self.temp_dir,file_new))
                 cnt+=1
@@ -105,7 +114,7 @@ class FileRename(object):
         print '%s files renamed' % cnt
     
 if __name__ == '__main__':
-    work_dir='/home/niedaocai/workspace/python/pythonDemo/src/files_rename'
+    #work_dir='/home/niedaocai/workspace/python/pythonDemo/src/files_rename'
     #work_dir='files_rename'
-    fr=FileRename(work_dir)
-    fr.rename_number(prefix='test_pic')
+    fr=FileRename('.')
+    fr.add_prefix(prefix='gk',before=False)
